# backend/contracts/openapi/control-plane.json
{
  "components": {
    "schemas": {
...
      "AccountStatus": {
        "additionalProperties": false,
        "properties": {
          "email_verified": {
            "title": "Email Verified",
            "type": "boolean"
          },
          "verification_required": {
            "title": "Verification Required",
            "type": "boolean"
          }
        },
        "required": [
          "email_verified",
          "verification_required"
        ],
        "title": "AccountStatus",
        "type": "object"
      },
...
      "CurrentUser": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email"
        ],
        "title": "CurrentUser",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceLimitCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ResourceLimitCode": {
        "enum": [
          "user_allowance",
          "global_capacity",
          "invalid_configuration"
        ],
        "title": "ResourceLimitCode",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "Polybot Control Plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/account": {
      "get": {
        "operationId": "accountStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests",
            "headers": {
              "Retry-After": {
                "description": "Seconds until another authentication attempt is allowed",
                "schema": {
                  "minimum": 1.0,
                  "type": "integer"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable",
            "headers": {
              "Retry-After": {
                "description": "Suggested seconds before retrying temporary resource admission",
                "schema": {
                  "minimum": 1.0,
                  "type": "integer"
                }
              }
            }
          }
        },
        "summary": "Account Status"
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "operationId": "currentUser",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests",
            "headers": {
              "Retry-After": {
                "description": "Suggested seconds before retrying temporary resource admission",
                "schema": {
                  "minimum": 1.0,
                  "type": "integer"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable",
            "headers": {
              "Retry-After": {
                "description": "Suggested seconds before retrying temporary resource admission",
                "schema": {
                  "minimum": 1.0,
                  "type": "integer"
                }
              }
            }
          }
        },
        "summary": "Me"
      }
    },
...
  }
}